Re #13: fix cyclic imports
