+++ mainWindow.py
@@ -23,15 +23,19 @@ from keySequences import KEY_SEQUENCES
 from model.commands.commands import AllCommands, BASIC_COMMAND_INFO
 from model.commands.parser import parseMCFunction
 from model.parsingUtils import Span
+from session.session import getSession, WindowId, saveSessionToFile
 from session.documents import Document, DocumentTypeDescription, getDocumentTypes, getErrorCounts
+from session import documentsImpl
 from gui.checkAllDialog import CheckAllDialog
 from gui.searchAllDialog import SearchAllDialog
 from gui.spotlightSearch import SpotlightSearchGui
 from model.pathUtils import FilePath
 from gui.datapackEditorGUI import DatapackEditorGUI
-from session.session import getSession, WindowId, saveSessionToFile
 from settings import applicationSettings
 from settings.settingsDialog import SettingsDialog
+
+
+documentsImpl.init()
 
 
 def frange(a: float, b: float, jump: float, *, includeLAst: bool = False):
