+++ session/documents.py
@@ -16,8 +16,6 @@ from Cat.utils import utils, Decorator
 from Cat.utils.profiling import logInfo, logError
 
 from Cat.utils.signals import CatSignal
-from model.commands.parser import parseMCFunction
-from model.commands.validator import checkMCFunction
 from model.pathUtils import fileNameFromFilePath, FilePath, getMTimeForFilePath, ZipFilePool, loadTextFile
 
 TTarget = TypeVar("TTarget")
@@ -419,65 +417,3 @@ class TextDocument(Document):
 
 	def fromRepr(self, string):
 		self.content = string
-
-
-@RegisterDocument('JSON', ext=['.json', '.mcmeta'], defaultLanguage='JSON')
-@RegisterContainer
-class JsonDocument(TextDocument):
-	"""docstring for Document"""
-	__slots__ = ()
-
-	def __typeCheckerInfo___(self):
-		# giving the type checker a helping hand...
-		super(JsonDocument, self).__typeCheckerInfo___()
-		self.filePath: FilePath = ''
-		self.filePathForDisplay: str = ''
-		self.fileName: str = ''
-		self.fileLocationAbsolute: FilePath = ''
-		self.fileLocationInProject: FilePath = ''
-		self.fileChanged: bool = False
-		self.documentChanged: bool = False
-		self.encoding: str = 'utf-8'
-		self.content: str = ''
-
-	filePath: FilePath = Serialized(default='', decorators=[
-		pd.FilePath(filters=[('JSON', '.json')])
-	])
-
-	encoding: str = Serialized(default='utf-8')
-
-	def validate(self) -> Sequence[Error]:
-		return []
-
-
-@RegisterDocument('mcFunction', ext=['.mcFunction'], defaultLanguage='MCFunction')
-@RegisterContainer
-class MCFunctionDocument(TextDocument):
-	"""docstring for Document"""
-	__slots__ = ()
-
-	def __typeCheckerInfo___(self):
-		# giving the type checker a helping hand...
-		super(MCFunctionDocument, self).__typeCheckerInfo___()
-		self.filePath: FilePath = ''
-		self.filePathForDisplay: str = ''
-		self.fileName: str = ''
-		self.fileLocationAbsolute: FilePath = ''
-		self.fileLocationInProject: FilePath = ''
-		self.fileChanged: bool = False
-		self.documentChanged: bool = False
-		self.encoding: str = 'utf-8'
-		self.content: str = ''
-
-	filePath: FilePath = Serialized(default='', decorators=[
-		pd.FilePath(filters=[('mcFunction', '.mcFunction')])
-	])
-
-	encoding: str = Serialized(default='utf-8')
-
-	def validate(self) -> Sequence[Error]:
-		tree, errors = parseMCFunction(self.content)
-		if tree is not None:
-			errors += checkMCFunction(tree)
-		return errors
-
